CResourceBrowser: Allow navigating directories with the Enter keys
Previously the resource browser required you to double click on
everything in order to interact with it (going into directories, or
launching handlers for particular files, etc).

This is kinda stinky workflow-wise long-term when flipping between
assets; so we can make the interface a little less mouse-centric by
allowing the use of the enter keys to also perform the same behavior.