CCamera: Allow toggleable camera movement speed
In a lot of 3D scene editors, it's quite common to have a default speed
that the camera can move at, plus a quicker one when shift (or some
other kind of keybind is held. We can replicate this here as a fixed
increment, but in the future this could also potentially be a
configurable UI setting.

Just makes the viewport expectations a little more natural like Blender
or 3DS Max.